Re-organizes files and creates welcome route
