+++ web/beacon-app/src/application/routes/paths.ts
@@ -5,6 +5,6 @@ export const routes = {
   support: '/support',
   profile: '/profile',
   dashboard: '/dashboard',
-  onboarding: '/onboarding',
+  welcome: '/welcome',
   setup: '/setup',
 };
