+++ web/beacon-app/src/components/layout/MainLayout.tsx
@@ -3,9 +3,9 @@ import { Outlet } from 'react-router-dom';
 
 import { LandingFooter } from '@/components/auth/LandingFooter';
 import { LandingHeader } from '@/components/auth/LandingHeader';
-import { WelcomePage } from '../ui/WelcomePage';
+import { WelcomePage } from '../../features/onboarding/components/Welcome';
 
-const MainLayout = () => {
+function MainLayout () {
   return (
     <Container>
       <LandingHeader />
