+++ web/beacon-app/src/features/onboarding/components/SetupTenant/CustomizeTenant.tsx
