+++ web/beacon-app/src/features/onboarding/components/SetupTenant/SetupTenant.tsx
