+++ web/beacon-app/src/features/onboarding/components/SetupTenant/TenantHeader.stories.tsx
@@ -2,7 +2,7 @@ import { Meta, Story } from '@storybook/react'
 import TenantHeader from './TenantHeader';
 
 export default {
-    title: 'ui/Welcome/TenantHeader',
+    title: 'onboarding/SetupTenant/TenantHeader',
     component: TenantHeader,
   } as Meta;
