+++ web/beacon-app/src/features/onboarding/components/SetupTenant/TenantHeader.tsx
