+++ web/beacon-app/src/features/onboarding/components/SetupTenant/index.ts
