+++ web/beacon-app/src/features/onboarding/components/Welcome/OnboardingHeader.stories.tsx
@@ -2,7 +2,7 @@ import { Meta, Story } from '@storybook/react'
 import OnboardingHeader from './OnboardingHeader';
 
 export default {
-    title: 'ui/Welcome/OnboardingHeader',
+    title: 'onboarding/Welcome/OnboardingHeader',
     component: OnboardingHeader,
   } as Meta;
