+++ web/beacon-app/src/features/onboarding/components/Welcome/OnboardingHeader.tsx
