+++ web/beacon-app/src/features/onboarding/components/Welcome/SetupYourTenant.tsx
