+++ web/beacon-app/src/features/onboarding/components/Welcome/ViewTutorials.tsx
