+++ web/beacon-app/src/features/onboarding/components/Welcome/WelcomePageLayout.tsx
@@ -1,0 +1,18 @@
+import { routes } from "@/application";
+import { Route } from "react-router-dom";
+import WelcomePage from "./WelcomePage";
+import MainLayout from "@/components/layout/MainLayout";
+
+export default function WelcomePageLayout () {
+    return (
+    <>
+    <Route path={routes.welcome} element={<MainLayout />}>
+        <Route
+          path={routes.welcome}
+          element={<WelcomePage />}
+        />
+    </Route>
+    </>
+    );
+  };
+
