+++ web/beacon-app/src/features/onboarding/components/Welcome/index.ts
